refactor(#120): move Expression Atlas constants out, drop dead registry paths
Moved EXPRESSION_ATLAS_BASE_URL and EXPRESSION_ATLAS_JSON_FILE_PATH
from core/constants.py to skills/expression_atlas/shared/constants.py.
Deleted REGISTRY_ROOT_PATH and TRANSCRIPTOMICS_DATASETS_PATH as dead
code — they had zero external consumers (orphan platform-registry
placeholders that were never wired up). Re-expressed the legacy JSON
path via parents[3] resolution (same final target). Dropped 4
module-import debug-log dumps. Updated 3 Python consumers and the
SKILL.md path reference. One of several PRs closing #120.
