+++ hvantk/skills/expression_atlas/shared/datasets.py
@@ -8,7 +8,7 @@ from urllib.parse import quote
 logger = logging.getLogger(__name__)
 
 from hvantk.core.utils.file_utils import download_file
-from hvantk.core.constants import EXPRESSION_ATLAS_BASE_URL
+from hvantk.skills.expression_atlas.shared.constants import EXPRESSION_ATLAS_BASE_URL
 
 
 @dataclass
